feat: create and list tournaments (#26)
* test: :white_check_mark: TDD E2E test for listing tournaments.

* feat: :sparkles: Implemented listing of tournaments.

* test: :white_check_mark: Added TDD E2E Testcase for creating a new scheduled tournament.

* feat: :construction: Initial setup of EditTournament(). Requires proper global state handling (must be implemented).

* feat: :sparkles: Added option to IdVersion to preset a id for a new object. save_tournament_base() uses this now.

* feat: :construction: Implemented initial TournamentEditorState to manage creating and editing tournaments.

* refactor: :construction: Refactoring EditTournament() to use TournamentEditorState. Next step is propper generic error handling.

* feat: :construction: Added global error handling and banner for blocking errors.

* feat: :construction: Adding toast manager and refactoring error handling in HomePage and EditTournament.

* feat: :construction: Implemented initial inputs for EditTournament.

* refactor: :construction: Fixing some bugs with reactive system.

Although finally fixing bug with ghost url of /sports. Problem was conditional renderiing of Outlet.

* feat: :construction: Added navigation links for stage configuration to EditTournament.

* test: :white_check_mark: Added hydration mark for e2e to make sure that tests start after full hydration.

* test: :white_check_mark: Adding requestAnimationFrame() in waitForAppHydration() for better interaction with webkit.

* chore: :green_heart: Updating to new playwright for better webkit performance.

* feat: :sparkles: Added Edit Tournament to List Tournaments.

+++ app_core/src/match_.rs
@@ -21,8 +21,8 @@ pub struct Match {
     group_id: Uuid,
     /// id of round
     round_id: Uuid,
-    /// number of match
-    number: usize,
+    /// number of match in round
+    number: u32,
     /// entrant a, either scheduled or concrete id
     side_a: ScheduledEntrant,
     /// entrant b, either scheduled or concrete id
